requirements.txt
# Core
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3

# Azure
azure-storage-blob==12.17.0
azure-identity==1.13.0

# Database
psycopg2-binary==2.9.6
sqlalchemy==2.0.20

# Data Processing
openpyxl==3.1.2
pyarrow==12.0.0

# Visualization
plotly==5.14.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
